+++ Products/CMFPlone/factory.py
@@ -5,7 +5,7 @@ from Products.GenericSetup.tool import SetupTool
 from plone.registry.interfaces import IRegistry
 from zope.component import queryUtility
 from zope.event import notify
-from zope.interface import implements
+from zope.interface import implementer
 from zope.site.hooks import setSite
 
 _TOOL_ID = 'portal_setup'
@@ -16,8 +16,8 @@ _CONTENT_PROFILE = 'plone.app.contenttypes:plone-content'
 _IMREALLYPLONE5 = True
 
 
+@implementer(INonInstallable)
 class HiddenProfiles(object):
-    implements(INonInstallable)
 
     def getNonInstallableProfiles(self):
         return [_DEFAULT_PROFILE,