+++ Products/CMFPlone/unicodeconflictresolver.py
@@ -1,15 +1,14 @@
-from zope.interface import implements
+from zope.interface import implementer
 from Products.PageTemplates.interfaces import IUnicodeEncodingConflictResolver
 
 from Products.CMFPlone.patches.unicodehacks import _unicode_replace
 
 
+@implementer(IUnicodeEncodingConflictResolver)
 class UTF8EncodingConflictResolver(object):
     """This resolver tries to decode a string from utf-8 and replaces it
        otherwise but logs a warning.
     """
-
-    implements(IUnicodeEncodingConflictResolver)
 
     def resolve(self, context, text, expression):
         return _unicode_replace(text)